+++ src/acetime/acetz.py
@@ -7,12 +7,11 @@ from datetime import datetime, tzinfo, timedelta, timezone
 
 from acetime.common import to_epoch_seconds
 from acetime.zone_processor import ZoneProcessor
-from acetime.zonedb_types import ZoneInfo, ZoneInfoMap
+from acetime.typing import ZoneInfo, ZoneInfoMap
 
 
 class acetz(tzinfo):
-    """An implementation of datetime.tzinfo using the ZoneProcessor class
-    from AceTime/tools.
+    """An implementation of datetime.tzinfo using the ZoneProcessor class.
     """
 
     def __init__(self, zone_info: ZoneInfo):