+++ tests/test_zone_processor.py
@@ -24,7 +24,7 @@ from acetime.zone_processor import MATCH_STATUS_PRIOR
 from acetime.zone_processor import MATCH_STATUS_EXACT_MATCH
 from acetime.zone_processor import MATCH_STATUS_WITHIN_MATCH
 from acetime.zone_processor import MATCH_STATUS_FAR_FUTURE
-from acetime.zonedb_types import ZoneEra
+from acetime.typing import ZoneEra
 
 
 class TestZoneProcessorHelperMethods(unittest.TestCase):