+++ CHANGELOG.md
@@ -7,6 +7,7 @@
       [dateutil.tz.gettz()](https://dateutil.readthedocs.io/en/stable/tz.html).
     * Package name refactoring
         * Rename `zonedb_types.py` to `typing.py`.
+        * Rename `acetime.acetz` (acetz.py) to `acetime.timezone` (timezone.py).
 * 0.6.1 (2023-04-01, TZDB 2023c)
     * Upgrade TZDB from 2023b to 2023c.
         * https://mm.icann.org/pipermail/tz-announce/2023-March/000079.html